requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.7
djangorestframework==3.14.0
flake8==6.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
